# sekyr-openapi.yml
openapi: 3.1.0
info:
  title: Beacon API
  description: Example API for managing beacons
  version: 1.0.0
components:
  schemas:
    TenantBaseModel:
      type: object
      properties:
        ID:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        CreatedAt:
          type: string
          format: date-time
          example: '2023-03-27T13:00:00Z'
        UpdatedAt:
          type: string
          format: date-time
          example: '2023-03-27T14:00:00Z'
        DeletedAt:
          type: object
          properties:
            Time:
              type: string
              format: date-time
              example: '2023-03-27T15:00:00Z'
            Deleted:
              type: boolean
              example: true
          required:
            - Time
            - Deleted
        TenantID:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
    Beacon:
      allOf:
        - $ref: '#/components/schemas/TenantBaseModel'
        - type: object
          properties:
            BeaconId:
              type: string
              format: uuid
              example: 123e4567-e89b-12d3-a456-426655440000
            GroupId:
              type: string
              format: uuid
              example: 123e4567-e89b-12d3-a456-426655440000
            OS:
              type: string
              example: Windows
            Arch:
              type: string
              example: x86_64
            Static:
              type: boolean
              example: true
            Upx:
              type: boolean
              example: true
            UpxLevel:
              type: integer
              example: 9
            Gzip:
              type: boolean
              example: true
            Transport:
              type: string
              example: http
            ReportAddr:
              type: string
              example: http://example.com/report
    Group:
      allOf:
        - $ref: '#/components/schemas/TenantBaseModel'
        - type: object
          properties:
            Name:
              type: string
              example: Test Group
            Description:
              type: string
              example: A group for testing
            Priority:
              type: integer
              example: 1
            Beacons:
              type: array
              items:
                $ref: '#/components/schemas/Beacon'
            Recipients:
              type: array
              items:
                $ref: '#/components/schemas/Recipient'
    Recipient:
      allOf:
        - $ref: '#/components/schemas/TenantBaseModel'
        - type: object
          properties:
            Name:
              type: string
              example: Recipient 1
            Type:
              type: string
              example: Email
            ContactInfo:
              type: string
              example: test@example.com
            Description:
              type: string
              example: A recipient for testing
            Groups:
              type: array
              items:
                $ref: '#/components/schemas/Group'
paths:
  /beacons:
    get:
      summary: List all beacons
      responses:
        '200':
          description: A list of beacons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Beacon'
    post:
      summary: Create a new beacon.
      tags:
        - beacons
      parameters:
        - name: ReportAddr
          in: query
          description: The URL of the report server.
          required: true
          schema:
            type: string
            example: http://example.com/report
        - name: OS
          in: query
          description: The operating system of the beacon.
          required: true
          schema:
            type: string
            example: Windows
        - name: Arch
          in: query
          description: The architecture of the beacon.
          required: true
          schema:
            type: string
            example: x86_64
        - name: BeaconId
          in: query
          description: The UUID of the beacon.
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426655440000
        - name: GroupId
          in: query
          description: The UUID of the group.
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426655440000
        - name: Static
          in: query
          description: Indicates if the beacon is static.
          schema:
            type: boolean
            default: false
            example: true
        - name: Upx
          in: query
          description: Indicates if the beacon is compressed using UPX.
          schema:
            type: boolean
            default: false
            example: true
        - name: UpxLevel
          in: query
          description: The compression level used by UPX.
          schema:
            type: integer
            default: 1
            example: 9
        - name: Gzip
          in: query
          description: Indicates if the beacon is compressed using Gzip.
          schema:
            type: boolean
            default: true
            example: true
        - name: Transport
          in: query
          description: The transport protocol used by the beacon.
          schema:
            type: string
            default: dns
            example: http
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: a beacon created from the submitted binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /beacons/{beaconId}:
    parameters:
      - in: path
        name: beaconId
        required: true
        schema:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
    get:
      summary: Get a specific beacon by ID
      responses:
        '200':
          description: The beacon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beacon'
    delete:
      summary: Delete a beacon by ID
      responses:
        '204':
          description: The beacon was deleted successfully
  /groups:
    get:
      summary: List all groups
      responses:
        '200':
          description: A list of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
    post:
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '201':
          description: The created group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}:
    parameters:
      - in: path
        name: groupId
        required: true
        schema:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
    get:
      summary: Get a specific group by ID
      responses:
        '200':
          description: The group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
    put:
      summary: Update an existing group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '200':
          description: The updated group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
    delete:
      summary: Delete a group by ID
      responses:
        '204':
          description: The group was deleted successfully
  /recipients:
    get:
      summary: List all recipients
      responses:
        '200':
          description: A list of recipients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipient'
    post:
      summary: Create a new recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipient'
      responses:
        '201':
          description: The created recipient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
  /recipients/{recipientId}:
    parameters:
      - in: path
        name: recipientId
        required: true
        schema:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
    get:
      summary: Get a specific recipient by ID
      responses:
        '200':
          description: The recipient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
    put:
      summary: Update an existing recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipient'
      responses:
        '200':
          description: The updated recipient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
    delete:
      summary: Delete a recipient by ID
      responses:
        '204':
          description: The recipient was deleted successfully